MFC (conceptually) r327377: Pull in strnlen for sys/boot environment
This is a direct commit to stable/11 due to restructuring of lib/libstand =>
stand/libsa in -HEAD. The diff remains the same and it is simply applied to
the previous location.

MFC r327377: Pull in strnlen from libc

strnlen is not used at the moment, but it will be when libfdt gets updated.
Prepare for the not-so-distant future by pulling in strnlen.


git-svn-id: https://svn.freebsd.org/base/stable/11@327957 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f