Record-only merges of recent loader fixes
Loader fixes were recently effectively MFC'd in r327942, r327944, r327945,
and r327957. These were all direct commits to stable/11 due to restructuring
in -HEAD where libstand and sys/boot are concerned, but achieve the same
effect as this range of commits.

Catch up on svn:mergeinfo following these, as this was missed in the
original commits to stable/11.


git-svn-id: https://svn.freebsd.org/base/stable/11@328105 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f