MFC r324431: patch(1): Don't overrun line buffer in some cases
Patches like file.txt attached to PR 190195 with a final line formed
like ">(EOL)" could cause a copy past the end of the current line buffer. In
the case of PR 191641, this caused a duplicate line to be copied into the
resulting file.

Instead of running past the end, treat it as if it were a blank line.

PR:		191641


git-svn-id: https://svn.freebsd.org/base/stable/11@328147 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f