MFC r328221:
Define a new __alloc_size2 attribute to complement the existing support.

At least on GCC7, calling __alloc_size(x) twice is not equivalent to
calling using the attribute once with two arguments. The later is the
documented use in GCC documentation so add a new alloc_size(n, x)
alternative to cover for the few places where it is used: basically:
calloc(3), reallocarray(3) and  mallocarray(9).

Submitted by:	Mark Millard


git-svn-id: https://svn.freebsd.org/base/stable/11@328308 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f