MFC r328287: Add /boot/overlays to runtime pkg, fix distrib-dirs METALOG
/boot/overlays was recently added without belonging to a package. It's only
used by bootloaders at the moment, so add it to the 'runtime' package to get
added with ubldr and friends.

Fix distrib-dirs METALOG generation while we're here. History elsewhere
seems to indicate that bapt@ fixed this to pull in all attributes from
mtrees while generating the METALOG. This fix got clobbered somewhere later,
so restore it.


git-svn-id: https://svn.freebsd.org/base/stable/11@328600 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f