MFC r320483 (by imp): Improve wdc error log pulling.
After review by the WDC engineers, improve how we pull down the
so-called 'e6' logs. The 'c6' logs are obsolete and support for them
has been removed because FreeBSD needed to pull them in chunks, which
is incompatible with the 0xc6 opcode implementation. Rather than leave
the code in place that produces bad log pulls, remove it.


git-svn-id: https://svn.freebsd.org/base/stable/11@328730 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f