Retire pnp.4th and the code needed only for 4th words used here.
This has never been installed. It was added to the tree disconnected
to the build in FreeBSD 5 (17 years ago) and has never been used as
far as I can tell. The desired improvements never really happened
(despite a couple minor cleanups along the way). It's relevance is
long past, so better to retire it.


git-svn-id: https://svn.freebsd.org/base/head@328779 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f